+++ CHANGELOG.md
@@ -4,8 +4,14 @@
 ### Added
 - Commands `pcs cluster node rename-corosync` and `pcs cluster node rename-cib`
   for cluster node renaming ([RHEL-149172])
+- `pcs constraint config` (and its variants for each constraint type) now list
+  resources in sets in the order defined in the CIB, instead of sorting them
+  alphabetically ([rhbz#2461143]) ([RHEL-176478])
+
 
 [RHEL-149172]: https://redhat.atlassian.net/browse/RHEL-149172
+[RHEL-176478]: https://redhat.atlassian.net/browse/RHEL-176478
+[rhbz#2461143]: https://bugzilla.redhat.com/show_bug.cgi?id=2461143
 
 ## [0.11.11] - 2026-01-06